+++ README.md
@@ -15,7 +15,7 @@ pip install -r requirements.txt
 ## Usage
 
 ```python
-from board import Board
+from sudoku_solver import Board
 
 # 0 represents an empty cell
 puzzle = [
