+++ sudoku_solver/__init__.py
@@ -1,0 +1,22 @@
+"""Sudoku solver package exporting the core Board class and helpers."""
+
+from .board import Board, find_determined_squares
+from .animation import (
+    animate_steps,
+    _format_grid,
+    GREEN,
+    RED,
+    YELLOW,
+)
+from . import example_boards
+
+__all__ = [
+    "Board",
+    "find_determined_squares",
+    "animate_steps",
+    "_format_grid",
+    "GREEN",
+    "RED",
+    "YELLOW",
+    "example_boards",
+]
